Add profile icon with working link
